Revert "sql: work around lack of runtime/trace in gccgo"
This reverts commit ce6c652e8f4355d5a8ee90aebdd7e129f1b53300.

gccgo now supports runtime/trace.
