--- pkg/sql/logictest/logic_gc.go
@@ -1,44 +1,0 @@
-// Copyright 2016 The Cockroach Authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
-// implied. See the License for the specific language governing
-// permissions and limitations under the License.
-
-// +build gc
-
-package logictest
-
-import (
-	"os"
-	"runtime/trace"
-)
-
-func (t *logicTest) traceStart(filename string) {
-	if t.traceFile != nil {
-		t.Fatalf("tracing already active")
-	}
-	var err error
-	t.traceFile, err = os.Create(filename)
-	if err != nil {
-		t.Fatalf("unable to open trace output file: %s", err)
-	}
-	if err := trace.Start(t.traceFile); err != nil {
-		t.Fatalf("unable to start tracing: %s", err)
-	}
-}
-
-func (t *logicTest) traceStop() {
-	if t.traceFile != nil {
-		trace.Stop()
-		t.traceFile.Close()
-		t.traceFile = nil
-	}
-}
